[refactor] extrac floor map as a component
